Optimize comment-display module
